requirements.txt
requests
requests-cache
lxml
progressbar
psycopg2
Django
django-tastypie
pytz
